Moved GenerateSecureRandomData() into helpers.go
